changed 3D noise format to float (GL16F)

+++ rendertools/include/noisetexture.h
@@ -38,7 +38,7 @@ public:
 private:
 	int m_edgeSize;
 
-	ManagedArray<uint8_t>	m_data;
+	ManagedArray<float>	m_data;
 
 	bool Allocate(int edgeSize);
